% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcroscope.R
\name{add_anatomy}
\alias{add_anatomy}
\title{Title}
\usage{
add_anatomy(plot, data = NULL, low = 700, high = 1400, guide = "none")
}
\arguments{
\item{plot}{}

\item{data}{}

\item{low}{}

\item{high}{}

\item{guide}{}
}
\description{
Title
}
