% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcroscope.R
\name{add_roi_overlay}
\alias{add_roi_overlay}
\title{Title}
\usage{
add_roi_overlay(
  plot,
  column,
  variable = "var",
  data = NULL,
  low = 2,
  high = 5,
  symmetric = T
)
}
\arguments{
\item{plot}{}

\item{column}{}

\item{variable}{}

\item{data}{}

\item{low}{}

\item{high}{}

\item{symmetric}{}
}
\description{
Title
}
