% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcroscope.R
\name{add_voxel_overlay}
\alias{add_voxel_overlay}
\title{Title}
\usage{
add_voxel_overlay(
  plot,
  data,
  name = "voxels",
  low = 2,
  high = 5,
  symmetric = T
)
}
\arguments{
\item{plot}{}

\item{data}{}

\item{low}{}

\item{high}{}

\item{symmetric}{}
}
\description{
Title
}
