% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{addtitle}
\alias{addtitle}
\title{Add a title to a slice series}
\usage{
addtitle(ssm, title, side = "right")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{title}{The text that will be used for the title.}

\item{side}{Which side of the slices to put the title on in row layout
(\code{draw(layout = "row")}): \code{"right"} (default) or \code{"left"}.
Ignored in column layout, where titles go above the slices.}
}
\value{
The slices series for continuation down the pipe.
}
\description{
Add a title to a slice series
}
\examples{
\dontrun{
sliceSeries(nrow = 8, begin=100, end=350) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  addtitle("Anatomy") \%>\%
  sliceSeries() \%>\%
  anatomy() \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  addtitle("Stats") \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
