% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{anatomy}
\alias{anatomy}
\title{Add anatomy slices}
\usage{
anatomy(ssm, volume = NULL, low = NULL, high = NULL,
  col = gray.colors(255, start = 0), alpha = NULL, name = "anatomy")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{volume}{3D matrix representing the volume from which to obtain slices.
Can be either numbers or hexadecimal colours}

\item{low}{Inclusive lower end of colour scale. Can be NULL if volume is a matrix of hexadecimals.}

\item{high}{Inclusive upper end of colour scale. Can be NULL if volume is a matrix of hexadecimals.}

\item{col}{The colour scale. Defaults to gray.}

\item{alpha}{Value between 0 and 1, 0 being full transparent and 1 fully opaque.}

\item{name}{Optional name.}
}
\value{
The slices series for continuation down the pipe.
}
\description{
Adds the anatomy slices, or anything that can serve as an underlay.
}
\examples{
\dontrun{
sliceSeries(nrow = 1, begin=200, end=300) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
