% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{contourSliceIndicator}
\alias{contourSliceIndicator}
\title{Add a slice indicator}
\usage{
contourSliceIndicator(ssm, volume, levels, dimension = NULL,
  slice = NULL, col = "red", lty = 1, lwd = 1,
  lineColour = "black")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{volume}{3D matrix representing the volume from which to draw the
contours.}

\item{levels}{Levels at which to draw contours - a single number or a vector.}

\item{dimension}{Dimension to use for contours. Will pick sensible
default if NULL.}

\item{slice}{Slice number to use for contours. Will pick sensible
default if NULL.}

\item{col}{The colour(s) for the contour(s). Defaults to red.}

\item{lty}{The line type(s) for the contour(s). Defaults to 1.}

\item{lwd}{The line width(s) for the contour(s). Defaults to 1.}

\item{lineColour}{The colour for the slice indicator lines. Defaults to black.}
}
\value{
The ssm augmented with a contour slice indication
}
\description{
Adds an indicator of the slices used in the slice series, displayed on top of
one or more contours.
}
\examples{
\dontrun{
sliceSeries(nrow=5, begin=150, end=250) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, 2, 6, symmetric = T, alpha=0.1) \%>\%
  legend("t-statistics") \%>\%
  contourSliceIndicator(anatVol, c(700, 1400)) \%>\%
  draw()
}
}
