% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagPoints.R
\name{convertPointToMRIcrotome}
\alias{convertPointToMRIcrotome}
\title{Title}
\usage{
convertPointToMRIcrotome(point, sliceList, assemblySlice)
}
\arguments{
\item{point}{a vector of length 3 in voxel coordinates}

\item{sliceList}{output of MRIcrotome}
}
\description{
Title
}
