% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{draw}
\alias{draw}
\title{Draw a slice series figure}
\usage{
draw(ssm, layout = "column")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{layout}{either "column" or "row" for how to layout the figure}
}
\value{
invisible NULL
}
\description{
Renders the slice series to the current graphics device. Opens a new page and
draws the assembled figure including all slice series, overlays, contours,
legends, and titles.
}
\examples{
\dontrun{
sliceSeries(nrow = 1, begin=200, end=300) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
