% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcroscope.R
\name{scale_fill_posneg}
\alias{scale_fill_posneg}
\title{Title}
\usage{
scale_fill_posneg(
  name = waiver(),
  ...,
  aesthetic = "fill",
  low = NULL,
  high = NULL
)
}
\arguments{
\item{...}{}

\item{aesthetic}{}

\item{low}{}

\item{high}{}
}
\description{
Title
}
