% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcroscope.R
\name{scale_fill_posneg_alpha}
\alias{scale_fill_posneg_alpha}
\title{Title}
\usage{
scale_fill_posneg_alpha(
  name = waiver(),
  ...,
  aesthetic = "fill",
  low = NULL,
  high = NULL,
  threshold = 1,
  lowalpha = 0.3
)
}
\arguments{
\item{...}{}

\item{aesthetic}{}

\item{low}{}

\item{high}{}
}
\description{
Title
}
