% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{sliceSeries}
\alias{sliceSeries}
\title{Initialize a slice series}
\usage{
sliceSeries(ssm = NULL, slices = NULL, nrow = NULL, ncol = NULL,
  dimension = NULL, begin = NULL, end = NULL)
}
\arguments{
\item{ssm}{Provided as part of an ongoing pipe, not usually specified by user}

\item{slices}{A vector of integers representing which slices to use. Can be
NULL, in which case nrow and ncol together with begin and end are used.}

\item{nrow}{Rows in the slice series. Set to 1 if NULL and ncol specified.}

\item{ncol}{Columns in the slice series. Set to 1 if NULL and nrow specified.}

\item{dimension}{Dimension as integer (1, 2, or 3) for slice series.}

\item{begin}{First slice. Can be NULL if slices explicitly specified.}

\item{end}{Last slice. Can be NULL if slices explicitly specified.}
}
\value{
Returns the slices series info for passing along the pipe.
}
\description{
Starts a new slices series, sets dimensions, slices to use, and rows and
columns.
}
\examples{
\dontrun{
sliceSeries(nrow = 1, begin=200, end=300) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
