% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{anatomySliceIndicator}
\alias{anatomySliceIndicator}
\title{Add a slice indicator}
\usage{
anatomySliceIndicator(
  ssm,
  volume,
  low,
  high,
  dimension = NULL,
  slice = NULL,
  col = gray.colors(255, start = 0),
  lineColour = "green"
)
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{volume}{3D matrix representing the volume from which to draw the
background slice.}

\item{low}{Inclusive low end of colour for background slice.}

\item{high}{Inclusive high end of colour for background slice.}

\item{dimension}{Dimension to use for background slice. Will pick sensible
default if NULL.}

\item{slice}{Slice number to use for background slice. Will pick sensible
default if NULL.}

\item{col}{Colour of the background slice. Defaults to gray.}

\item{lineColour}{The colour of the slice indicator lines. Defaults to green.}
}
\value{
The ssm object augmented with a sliceIndicator
}
\description{
Adds an indicator of the slices used in the slice series, displayed on top of
an anatomy slice.
}
\examples{
\dontrun{
sliceSeries(nrow=5, begin=150, end=250) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, 2, 6, symmetric = T, alpha=0.1) \%>\%
  legend("t-statistics") \%>\%
  anatomySliceIndicator(anatVol, 700, 1400) \%>\%
  draw()
}
}
