% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{contours}
\alias{contours}
\title{Add contours to a slice series}
\usage{
contours(ssm, volume, levels, col = "red", lty = 1, lwd = 1,
  name = "contours")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{volume}{3D matrix representing the volume from which to obtain contours}

\item{levels}{A vector of levels at which to draw the contours.}

\item{col}{The colour for the contours. Can be a single value, in which case
all levels will use the same colour, or a vector if different colours for
different levels are desired. Default is red.}

\item{lty}{The line type for the contours. Can be a single value, in which case
all levels will use the same line type, or a vector if different line types for
different levels are desired. Default is 1.}

\item{lwd}{The line width for the contours. Can be a single value, in which case
all levels will use the same line width, or a vector if different line widths for
different levels are desired. Default is 1.}

\item{name}{An optional name.}
}
\value{
The slices series for continuation down the pipe.
}
\description{
Adds contours to every slice in the series, with control over the contour
levels and attributes.
}
\examples{
\dontrun{
sliceSeries(nrow = 1, begin=200, end=300) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  legend("t-statistics") \%>\%
  contours(abs(stats), levels=c(3,5), lwd=2, lty=c(3,1), col="green") \%>\%
  draw()
}
}
