% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{draw}
\alias{draw}
\title{Title}
\usage{
draw(ssm, layout = "column")
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{layout}{either "column" or "row" for how to layout the figure}
}
\value{
invisible NULL
}
\description{
Title
}
