% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{grobify}
\alias{grobify}
\title{Turn the slice series into grobs.}
\usage{
grobify(ssm, layout = "column", titlePars = gpar(),
  legendPars = gpar(), bgCol = NULL)
}
\arguments{
\item{ssm}{The slice series}

\item{layout}{Whether to lay out by column (the default) or by row.}

\item{titlePars}{A \code{gpar} list configuring text grobs when generating titles.
useful for setting e.g. \code{fontsize} and \code{col}}

\item{legendPars}{A \code{gpar} list configuring text grobs when generating legends
useful for setting e.g. \code{fontsize} and \code{col}}

\item{bgCol}{Either NULL or a string indicating what colour the figure background
should be.}
}
\value{
A gTree object with the figure, to be drawn with \link{grid.draw}
}
\description{
Takes the slice series, creates grobs where necessary, and returns a gTree.
Can then be incorporated into other grid functions.
}
\examples{
\dontrun{
sliceSeriesGridTree <- sliceSeries(nrow=5, begin=150, end=250) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, 2, 6, symmetric = T, alpha=0.1) \%>\%
  legend("t-statistics") \%>\%
  anatomySliceIndicator(anatVol, 700, 1400) \%>\%
  grobify()
}
}
