% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{legend}
\alias{legend}
\title{Add a legend}
\usage{
legend(ssm, description = NULL)
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{description}{The text that will be used to describe the legend. Can be
NULL, in which case the colour bar is shown without a description.}
}
\value{
The slices series for continuation down the pipe.
}
\description{
Adds a legend based on the current slice series. Colour bars, contours, etc.,
are taken from the immediately preceding element.
}
\examples{
\dontrun{
sliceSeries(nrow = 8, begin=100, end=350) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  addtitle("Anatomy") \%>\%
  sliceSeries() \%>\%
  anatomy() \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  addtitle("Stats") \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
