% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIcrotome-package.r, R/ggMRIcrotome.R
\docType{package}
\name{MRIcrotome}
\alias{MRIcrotome-package}
\alias{MRIcrotome}
\title{MRIcrotome.}
\usage{
MRIcrotome(
  anatomy,
  labels,
  labelDefs,
  sliceList,
  assembleDir = "Y",
  centreSlices = T,
  shrinkToPols = T,
  sliceOffset = 0,
  smooth = T,
  smoothness = 1.5,
  simplify = F,
  ...
)
}
\arguments{
\item{anatomy}{}

\item{labels}{}

\item{labelDefs}{}

\item{sliceList}{}

\item{assembleDir}{}

\item{centreSlices}{}

\item{shrinkToPols}{}

\item{sliceOffset}{}
}
\description{
Functions to create publication-ready figures from 3D volumes.
}
