% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeFuncs.R
\name{overlay}
\alias{overlay}
\title{Add overlay slices}
\usage{
overlay(
  ssm,
  volume,
  low = NULL,
  high = NULL,
  col = defaultCol(),
  symmetric = FALSE,
  rCol = defaultRCol(),
  alpha = NULL,
  underTransparent = TRUE,
  name = NULL,
  box = FALSE,
  interpolate = TRUE
)
}
\arguments{
\item{ssm}{The slice series info, usually passed along the pipe and specified
by the user}

\item{volume}{3D matrix representing the volume from which to obtain slices.
Can be either numbers or hexadecimal colours}

\item{low}{Inclusive lower end of colour scale. Can be NULL if volume is a matrix of
hexadecimals. If symmetric=TRUE, then treat this as an absolute value (i.e.
2 would be >2 and < -2)}

\item{high}{Inclusive upper end of colour scale.}

\item{col}{The colour scale. See defaultCol for default.}

\item{symmetric}{Whether the colour scale is symmetric.}

\item{rCol}{The reverse colour scale. Used if symmetric=TRUE. See defaultRCol
for default.}

\item{alpha}{Value between 0 and 1, 0 being full transparent and 1 fully opaque.}

\item{underTransparent}{Whether to make the under colour transparent. Defaults to TRUE.}

\item{name}{Optional name.}

\item{box}{Whether to draw a box around the overlay. Defaults to FALSE.}

\item{interpolate}{Logical; whether to apply linear interpolation to the
overlay image when drawing. Set to \code{FALSE} for sharp boundaries
(e.g., label overlays with discrete colours). Defaults to TRUE.}
}
\value{
The slices series for continuation down the pipe.
}
\description{
Adds overlay slices; usually used for statistics, but can be anything.
Differs from anatomy in that the background is transparent.
}
\examples{
\dontrun{
sliceSeries(nrow = 1, begin=200, end=300) \%>\%
  anatomy(anatVol, low=700, high=1400) \%>\%
  overlay(stats, low=2, high=6, symmetric = T) \%>\%
  legend("t-statistics") \%>\%
  draw()
}
}
